refactor: move tags to his own package
To be able to consume tags files from any worker with the same values as
yggdrasil.
