+++ pkg/tags/tags_test.go
@@ -1,4 +1,4 @@
-package main
+package tags
 
 import (
 	"io"
